requirements.txt
PyMuPDF
Flask==3.0.3
img2table==1.2.11
llama-index==0.9.25.post1
llama_cpp_python
numpy==1.26.4
paddlepaddle
paddleocr
pandas==2.2.2
pdf2image==1.17.0
Pillow==10.3.0
Requests==2.31.0
streamlit==1.33.0
transformers==4.35.2
torch
torchvision
pypdf
fastapi
uvicorn[standard]
python-dotenv
pymongo
sentence-transformers